multiple networks logic, feedback applied
